Add spacy NER service option
