+++ docker-compose.dev.yml
@@ -56,6 +56,7 @@ services:
       - rabbitmq
       - convert-document
       - extract-polyglot
+      - extract-spacy
     tmpfs: /tmp
     volumes:
       - archive-data:/data
@@ -86,6 +87,7 @@ services:
       - rabbitmq
       - convert-document
       - extract-polyglot
+      - extract-spacy
     tmpfs: /tmp
     volumes:
       - archive-data:/data
